ingestion/producer/requirements.txt
confluent-kafka[avro,schemaregistry]>=2.3.0
psycopg2-binary>=2.9.9
Faker>=24.0.0
